src/experience/requirements.txt
coverage>=4.5.2
requests>=2.21.0
django-cors-headers>=2.4.0 

